No Ticket: Ignore extra keys sent to envelop and which not present on the actual document. (pull request #4)
Issue Origin: Single singer ESS we are sending the "owner_initial" anchor which was not present on template, This will be present only on the Improvements addendum.

Fix: Added the generic fix to ignore all the tags which was not present on document.

Reference email: DocuSign Support Case: 04744847 "DocuSign - Templates"

Approved-by: Krishna Mohan <krishna.mohan@sumerusolutions.com>